+++ index.js
@@ -3,14 +3,16 @@ const fs = require('fs');
 const Manager = require('./lib/Manager.js')
 const Engineer = require('./lib/Engineer.js')
 const Intern = require('./lib/Intern.js')
-const Employee = require('./lib/Employee.js');
-const { start } = require('repl');
+const startHtml = require('./src/starthtml.js')
+const addHtml = require('./src//addhtml.js')
+const finishHtml = require('./src/finishtml.js')
+
 
 const employees = [];
 
 function initialize() {
   startHtml();
-  addMember();
+  chooseProfile();
 }
 
 
@@ -94,109 +96,7 @@ function chooseProfile() {
     });
 });
 }
-
-function startHtml() {
-  const html = `<html lang="en">
-  <head>
-      <meta charset="UTF-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
-      <meta http-equiv="X-UA-Compatible" content="ie=edge">
-      <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-      <title>Team Profile</title>
-  </head>
-  <body>
-      <nav class="navbar navbar-dark bg-danger mb-5">
-          <span class="navbar-brand mb-0 h1 w-100 text-center">Team Profile</span>
-      </nav>
-      <div class="container">`;
-  fs.writeFile("./dist/team-profile.html", html, function(err) {
-      if (err) {
-          console.log(err);
-      }
-  });
-  console.log("start");
-}
-
-function addHtml(member) {
-  return new Promise(function(resolve, reject) {
-      const name = member.getName();
-      const role = member.getRole();
-      const id = member.getId();
-      const email = member.getEmail();
-      let data = "";
-      if (role === "Engineer") {
-          const gitHub = member.getGithub();
-          data = `<div class="col-6">
-          <div class="card mx-auto mb-3" style="width: 18rem">
-          <h5 class="card-header">${name}<br /><br />Engineer</h5>
-          <ul class="list-group list-group-flush">
-              <li class="list-group-item">ID: ${id}</li>
-              <li class="list-group-item">Email Address: ${email}</li>
-              <li class="list-group-item">GitHub: ${gitHub}</li>
-          </ul>
-          </div>
-      </div>`;
-      } else if (role === "Intern") {
-          const school = member.getSchool();
-          data = `<div class="col-6">
-          <div class="card mx-auto mb-3" style="width: 18rem">
-          <h5 class="card-header">${name}<br /><br />Intern</h5>
-          <ul class="list-group list-group-flush">
-              <li class="list-group-item">ID: ${id}</li>
-              <li class="list-group-item">Email Address: ${email}</li>
-              <li class="list-group-item">School: ${school}</li>
-          </ul>
-          </div>
-      </div>`;
-      }else if (role === "Manager") {
-        const officePhone = member.getOfficeNumber();
-          data = `<div class="col-6">
-          <div class="card mx-auto mb-3" style="width: 18rem">
-          <h5 class="card-header">${name}<br /><br />Manager</h5>
-          <ul class="list-group list-group-flush">
-              <li class="list-group-item">ID: ${id}</li>
-              <li class="list-group-item">Email Address: ${email}</li>
-              <li class="list-group-item">Office Phone: ${officePhone}</li>
-          </ul>
-          </div>
-      </div>`
-      } else {
-        data = `<div class="col-6">
-        <div class="card mx-auto mb-3" style="width: 18rem">
-        <h5 class="card-header">${name}<br /><br />Intern</h5>
-        <ul class="list-group list-group-flush">
-            <li class="list-group-item">ID: ${id}</li>
-            <li class="list-group-item">Email Address: ${email}</li>
-        </ul>
-        </div>
-    </div>`;
-      }
-      console.log("adding team member");
-      fs.appendFile("./output/team-profile.html", data, function (err) {
-          if (err) {
-              return reject(err);
-          };
-          return resolve();
-      });
-  });
-
-}
-
-function finishHtml() {
-    const html = ` </div>
-    </div>
-    
-</body>
-</html>`;
-
-    fs.appendFile("./dist/team-profile.html", html, function (err) {
-        if (err) {
-            console.log(err);
-        };
-    });
-    console.log("end");
-}
-initApp();
+initialize();
 
 
 
@@ -245,3 +145,107 @@ initApp();
 //   </header>
 // </body>
 // </html>`;
+
+
+
+// function startHtml() {
+//   const html = `<html lang="en">
+//   <head>
+//       <meta charset="UTF-8">
+//       <meta name="viewport" content="width=device-width, initial-scale=1.0">
+//       <meta http-equiv="X-UA-Compatible" content="ie=edge">
+//       <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+//       <title>Team Profile</title>
+//   </head>
+//   <body>
+//       <nav class="navbar navbar-dark bg-danger mb-5">
+//           <span class="navbar-brand mb-0 h1 w-100 text-center">Team Profile</span>
+//       </nav>
+//       <div class="container">`;
+//   fs.writeFile("./dist/team-profile.html", html, function(err) {
+//       if (err) {
+//           console.log(err);
+//       }
+//   });
+//   console.log("start");
+// }
+
+// function addHtml(member) {
+//   return new Promise(function(resolve, reject) {
+//       const name = member.getName();
+//       const role = member.getRole();
+//       const id = member.getId();
+//       const email = member.getEmail();
+//       let data = "";
+//       if (role === "Engineer") {
+//           const gitHub = member.getGithub();
+//           data = `<div class="col-6">
+//           <div class="card mx-auto mb-3" style="width: 18rem">
+//           <h5 class="card-header">${name}<br /><br />Engineer</h5>
+//           <ul class="list-group list-group-flush">
+//               <li class="list-group-item">ID: ${id}</li>
+//               <li class="list-group-item">Email Address: ${email}</li>
+//               <li class="list-group-item">GitHub: ${gitHub}</li>
+//           </ul>
+//           </div>
+//       </div>`;
+//       } else if (role === "Intern") {
+//           const school = member.getSchool();
+//           data = `<div class="col-6">
+//           <div class="card mx-auto mb-3" style="width: 18rem">
+//           <h5 class="card-header">${name}<br /><br />Intern</h5>
+//           <ul class="list-group list-group-flush">
+//               <li class="list-group-item">ID: ${id}</li>
+//               <li class="list-group-item">Email Address: ${email}</li>
+//               <li class="list-group-item">School: ${school}</li>
+//           </ul>
+//           </div>
+//       </div>`;
+//       }else if (role === "Manager") {
+//         const officePhone = member.getOfficeNumber();
+//           data = `<div class="col-6">
+//           <div class="card mx-auto mb-3" style="width: 18rem">
+//           <h5 class="card-header">${name}<br /><br />Manager</h5>
+//           <ul class="list-group list-group-flush">
+//               <li class="list-group-item">ID: ${id}</li>
+//               <li class="list-group-item">Email Address: ${email}</li>
+//               <li class="list-group-item">Office Phone: ${officePhone}</li>
+//           </ul>
+//           </div>
+//       </div>`
+//       } else {
+//         data = `<div class="col-6">
+//         <div class="card mx-auto mb-3" style="width: 18rem">
+//         <h5 class="card-header">${name}<br /><br />Intern</h5>
+//         <ul class="list-group list-group-flush">
+//             <li class="list-group-item">ID: ${id}</li>
+//             <li class="list-group-item">Email Address: ${email}</li>
+//         </ul>
+//         </div>
+//     </div>`;
+//       }
+//       console.log("adding team member");
+//       fs.appendFile("./output/team-profile.html", data, function (err) {
+//           if (err) {
+//               return reject(err);
+//           };
+//           return resolve();
+//       });
+//   });
+
+// }
+
+// function finishHtml() {
+//     const html = ` </div>
+//     </div>
+    
+// </body>
+// </html>`;
+
+//     fs.appendFile("./dist/team-profile.html", html, function (err) {
+//         if (err) {
+//             console.log(err);
+//         };
+//     });
+//     console.log("end");
+// }
