+++ src/finishtml.js
@@ -1,0 +1,16 @@
+function finishHtml() {
+    const html = ` </div>
+    </div>
+    
+</body>
+</html>`;
+
+    fs.appendFile("./dist/team-profile.html", html, function (err) {
+        if (err) {
+            console.log(err);
+        };
+    });
+    console.log("end");
+}
+
+module.exports = finishHtml;
